requirements.txt
python-decouple
Flask
mysql-connector-python
python-dotenv 
flask-mail
flask-bcrypt
PyJWT
requests